+++ src/gateways/PaymentIntents.php
@@ -28,17 +28,21 @@ use craft\commerce\stripe\Plugin as StripePlugin;
 use craft\commerce\stripe\responses\PaymentIntentResponse;
 use craft\commerce\stripe\web\assets\intentsform\IntentsFormAsset;
 use craft\elements\User;
+use craft\errors\ElementNotFoundException;
 use craft\helpers\Json;
 use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
 use craft\web\View;
+use Exception;
 use Stripe\Card;
 use Stripe\Charge as StripeCharge;
 use Stripe\PaymentIntent;
 use Stripe\PaymentMethod;
 use Stripe\Refund;
 use Stripe\Subscription as StripeSubscription;
+use Throwable;
 use yii\base\NotSupportedException;
+use function count;
 
 /**
  * This class represents the Stripe Payment Intents gateway
@@ -127,12 +131,11 @@ class PaymentIntents extends BaseGateway
     {
         $this->configureStripeClient();
         try {
-            /** @var PaymentIntent $intent */
             $intent = PaymentIntent::retrieve($reference);
             $intent->capture([], ['idempotency_key' => $reference]);
 
             return $this->createPaymentResponseFromApiResource($intent);
-        } catch (\Exception $exception) {
+        } catch (Exception $exception) {
             return $this->createPaymentResponseFromError($exception);
         }
     }
@@ -180,7 +183,7 @@ class PaymentIntents extends BaseGateway
         if (!empty($intentData['payment_method'])) {
             try {
                 $this->_confirmPaymentIntent($stripePaymentIntent, $transaction);
-            } catch (\Exception $exception) {
+            } catch (Exception $exception) {
                 return $this->createPaymentResponseFromError($exception);
             }
         }
@@ -219,7 +222,7 @@ class PaymentIntents extends BaseGateway
                 $refund = Refund::create($request);
 
                 return $this->createPaymentResponseFromApiResource($refund);
-            } catch (\Exception $exception) {
+            } catch (Exception $exception) {
                 return $this->createPaymentResponseFromError($exception);
             }
         }
@@ -248,7 +251,7 @@ class PaymentIntents extends BaseGateway
             }
 
             return $this->createPaymentResponseFromApiResource($refund);
-        } catch (\Exception $exception) {
+        } catch (Exception $exception) {
             return $this->createPaymentResponseFromError($exception);
         }
     }
@@ -279,16 +282,14 @@ class PaymentIntents extends BaseGateway
                     $description = $stripeResponse->type;
             }
 
-            $paymentSource = new PaymentSource([
+            return new PaymentSource([
                 'userId' => $userId,
                 'gatewayId' => $this->id,
                 'token' => $stripeResponse->id,
                 'response' => $stripeResponse->toArray(),
                 'description' => $description,
             ]);
-
-            return $paymentSource;
-        } catch (\Throwable $exception) {
+        } catch (Throwable $exception) {
             throw new PaymentSourceException($exception->getMessage());
         }
     }
@@ -304,7 +305,7 @@ class PaymentIntents extends BaseGateway
         $customer = StripePlugin::getInstance()->getCustomers()->getCustomer($this->id, $user);
         $paymentMethods = PaymentMethod::all(['customer' => $customer->reference, 'type' => 'card']);
 
-        if (\count($paymentMethods->data) === 0) {
+        if (count($paymentMethods->data) === 0) {
             throw new PaymentSourceException(Craft::t('commerce-stripe', 'No payment sources are saved to use for subscriptions.'));
         }
 
@@ -316,7 +317,7 @@ class PaymentIntents extends BaseGateway
         if ($parameters->trialDays !== null) {
             $subscriptionParameters['trial_period_days'] = (int)$parameters->trialDays;
         } else if ($parameters->trialEnd !== null) {
-            $subscriptionParameters['trial_end'] = (int)$parameters->trialEnd;;
+            $subscriptionParameters['trial_end'] = (int)$parameters->trialEnd;
         } else {
             $subscriptionParameters['trial_from_plan'] = true;
         }
@@ -331,7 +332,7 @@ class PaymentIntents extends BaseGateway
 
         try {
             $subscription = StripeSubscription::create($event->parameters);
-        } catch (\Throwable $exception) {
+        } catch (Throwable $exception) {
             Craft::warning($exception->getMessage(), 'stripe');
 
             throw new SubscriptionException(Craft::t('commerce-stripe', 'Unable to subscribe at this time.'));
@@ -347,10 +348,9 @@ class PaymentIntents extends BaseGateway
     {
         $this->configureStripeClient();
         try {
-            /** @var PaymentMethod $paymentMethod */
             $paymentMethod = PaymentMethod::retrieve($token);
             $paymentMethod->detach();
-        } catch (\Throwable $throwable) {
+        } catch (Throwable $throwable) {
             // Assume deleted.
         }
 
@@ -433,8 +433,8 @@ class PaymentIntents extends BaseGateway
      * Handle a failed invoice by updating the subscription data for the subscription it failed.
      *
      * @param array $data
-     * @throws \Throwable
-     * @throws \craft\errors\ElementNotFoundException
+     * @throws Throwable
+     * @throws ElementNotFoundException
      * @throws \yii\base\Exception
      */
     protected function handleInvoiceFailed(array $data): void
@@ -539,7 +539,7 @@ class PaymentIntents extends BaseGateway
             $this->_confirmPaymentIntent($stripePaymentIntent, $transaction);
 
             return $this->createPaymentResponseFromApiResource($stripePaymentIntent);
-        } catch (\Exception $exception) {
+        } catch (Exception $exception) {
             return $this->createPaymentResponseFromError($exception);
         }
     }