+++ src/models/Customer.php
@@ -14,6 +14,7 @@ use craft\commerce\Plugin as Commerce;
 use craft\commerce\stripe\records\Customer as CustomerRecord;
 use craft\elements\User;
 use yii\base\InvalidConfigException;
+use craft\validators\UniqueValidator;
 
 /**
  * Stripe customer model
@@ -103,11 +104,12 @@ class Customer extends Model
     /**
      * @inheritdoc
      */
-    public function rules(): array
+    protected function defineRules(): array
     {
-        return [
-            [['reference'], 'unique', 'targetAttribute' => ['gatewayId', 'reference'], 'targetClass' => CustomerRecord::class],
-            [['gatewayId', 'userId', 'reference', 'response'], 'required'],
-        ];
+        $rules = parent::defineRules();
+        $rules[] = [['reference'], UniqueValidator::class, 'targetClass' => CustomerRecord::class];
+        $rules[] = [['gatewayId', 'userId', 'reference', 'response'], 'required'];
+
+        return $rules;
     }
 }