+++ src/models/PaymentIntent.php
@@ -14,8 +14,9 @@ use craft\commerce\elements\Order;
 use craft\commerce\models\Transaction;
 use craft\commerce\Plugin as Commerce;
 use craft\commerce\stripe\Plugin as StripePlugin;
-use craft\commerce\stripe\records\Customer as CustomerRecord;
+use craft\commerce\stripe\records\PaymentIntent as PaymentIntentRecord;
 use craft\elements\User;
+use craft\validators\UniqueValidator;
 
 /**
  * Stripe Payment Intent model
@@ -167,7 +168,7 @@ class PaymentIntent extends Model
      */
     public function getTransaction(): ?Transaction
     {
-        if (null === $this->_tranasction) {
+        if (null === $this->_transaction) {
             $this->_transaction = Commerce::getInstance()->getTransactions()->getTransactionByHash($this->transactionHash);
         }
 
@@ -177,11 +178,13 @@ class PaymentIntent extends Model
     /**
      * @inheritdoc
      */
-    public function rules(): array
+    protected function defineRules(): array
     {
-        return [
-            [['reference'], 'unique', 'targetAttribute' => ['gatewayId', 'reference'], 'targetClass' => CustomerRecord::class],
-            [['gatewayId', 'customerId', 'reference', 'intentData', 'orderId', 'transactionHash'], 'required'],
-        ];
+        $rules = parent::defineRules();
+
+        $rules[] = [['reference'], UniqueValidator::class, 'targetClass' => PaymentIntentRecord::class];
+        $rules[] = [['gatewayId', 'customerId', 'reference', 'intentData', 'orderId', 'transactionHash'], 'required'];
+
+        return $rules;
     }
 }