+++ src/models/forms/payment/PaymentIntent.php
@@ -32,9 +32,11 @@ class PaymentIntent extends CreditCardPaymentForm
     /**
      * @inheritdoc
      */
-    public function rules(): array
+    protected function defineRules(): array
     {
-        return [[['paymentMethodId'], 'required']];
+        return [
+            [['paymentMethodId'], 'required'],
+        ];
     }
 
     /**