+++ src/responses/ChargeResponse.php
@@ -9,6 +9,7 @@ namespace craft\commerce\stripe\responses;
 
 use craft\commerce\base\RequestResponseInterface;
 use craft\commerce\errors\NotImplementedException;
+use Stripe\Refund;
 
 class ChargeResponse implements RequestResponseInterface
 {
@@ -62,6 +63,12 @@ class ChargeResponse implements RequestResponseInterface
      */
     public function isSuccessful(): bool
     {
+        if (array_key_exists('object', $this->data) && $this->data['object'] == Refund::OBJECT_NAME) {
+            if (array_key_exists('status', $this->data) && in_array($this->data['status'], [Refund::STATUS_PENDING], true)) {
+                return false;
+            }
+        }
+
         return array_key_exists('status', $this->data) && $this->data['status'] === 'succeeded';
     }
 
@@ -70,6 +77,12 @@ class ChargeResponse implements RequestResponseInterface
      */
     public function isProcessing(): bool
     {
+        if (array_key_exists('object', $this->data) && $this->data['object'] == Refund::OBJECT_NAME) {
+            if (array_key_exists('status', $this->data) && in_array($this->data['status'], [Refund::STATUS_PENDING], true)) {
+                return true;
+            }
+        }
+
         return $this->_processing;
     }
 
@@ -142,6 +155,12 @@ class ChargeResponse implements RequestResponseInterface
      */
     public function getMessage(): string
     {
+        if (array_key_exists('object', $this->data) && $this->data['object'] == Refund::OBJECT_NAME) {
+            if (array_key_exists('status', $this->data) && in_array($this->data['status'], [Refund::STATUS_PENDING], true)) {
+                return 'Refund processing';
+            }
+        }
+
         if (empty($this->data['message'])) {
             return '';
         }